Code reorganisation for next step (real trace transformer)
